requirements.txt
flaskflask-corsrequestsgunicorn